feat(entidades categoria) read, create, update, categoriaById, asignar jogador categoria

+++ src/jogadores/jogadores.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, Logger } from '@nestjs/common';
+import {
+  Injectable,
+  Logger,
+  BadRequestException,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CriarJogadorDto } from './dto/criar-jogador.dto';
@@ -27,9 +32,15 @@ export class JogadoresService {
     return jogador;
   }
 
-  async criarAtualizarJugador(
-    criarJogadorDto: CriarJogadorDto,
-  ): Promise<Jogador> {
+  async criarJugador(criarJogadorDto: CriarJogadorDto): Promise<Jogador> {
+    const { email } = criarJogadorDto;
+    const jogador = await this.jogadorModel.findOne({ email });
+
+    if (jogador) {
+      throw new BadRequestException(
+        `Jugador con email ${email} ya esta registrado`,
+      );
+    }
     const newJogador = new this.jogadorModel(criarJogadorDto);
     return await newJogador.save();
   }
@@ -38,6 +49,19 @@ export class JogadoresService {
     jogadorID: string,
     criarJogadorDto: CriarJogadorDto,
   ): Promise<Jogador> {
+    const [jogador, jogadorByEmail] = await Promise.all([
+      this.jogadorModel.findById(jogadorID),
+      this.jogadorModel.findOne({ email: criarJogadorDto.email }),
+    ]);
+
+    if (!jogador) {
+      throw new NotFoundException(`Jogador con ID ${jogadorID} no existe`);
+    }
+
+    if (jogadorByEmail && jogadorByEmail._id !== jogadorID) {
+      throw new BadRequestException(`Email ${jogadorByEmail.email} ya existe`);
+    }
+
     const jogadorCurr = await this.jogadorModel.findByIdAndUpdate(
       jogadorID,
       criarJogadorDto,
@@ -48,6 +72,10 @@ export class JogadoresService {
   }
 
   async deleteJogador(jogadorID: string) {
+    const jogador = await this.jogadorModel.findById(jogadorID);
+    if (!jogador) {
+      throw new NotFoundException(`Jogador con ID ${jogadorID} no existe`);
+    }
     const deleteJogador = await this.jogadorModel.findByIdAndDelete(jogadorID);
     return deleteJogador;
   }
